requirements.txt
Django==4.2.7
Pillow
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.6
whitenoise==6.6.0
gunicorn==21.2.0
django-debug-toolbar==4.2.0
python-dateutil==2.8.2
pytz==2023.3
django-storages==1.14.2
boto3>=1.26.0  # For S3 if needed
django-ratelimit
django-crispy-forms==2.1
crispy-bootstrap5==0.7
whitenoise==6.6.0